split graph implementation into cpp file
